ENH: move writer out of test
