+++ redsky/savers.py
@@ -15,47 +15,40 @@
 ##############################################################################
 
 import numpy as np
-from uuid import uuid4
+import uuid
 import os
 
 
-class NPYSaver:
+class NpyWriter:
     """
-    Instantiating this class creates a new resource. Writing adds datums to
-    that resource.
+    Each call to the ``write`` method saves a file and creates a new filestore
+    resource and datum record.
     """
 
     SPEC = 'npy'
-    EXT = '.npy'
 
     def __init__(self, fs, root):
         self._root = root
         self._closed = False
         self._fs = fs
-        self._fp = os.path.join(self._root, str(uuid4()))
-        self._resource = self._fs.insert_resource(self.SPEC,
-                                                  self._fp + self.EXT,
-                                                  resource_kwargs={})
         # Open and stash a file handle (e.g., h5py.File) if applicable.
 
     def write(self, data):
         """
-        Save data to file, generate new datum_id, insert datum, return
-        datum_id.
+        Save data to file, generate and insert new resource and datum.
         """
         if self._closed:
-            raise RuntimeError('new_resource must be called first')
-        np.save(self._fp, data)
-        # save the data to self._fp
-        datum_id = str(uuid4())
-        self._fs.insert_datum(resource=self._resource, datum_id=datum_id,
+            raise RuntimeError('This writer has been closed.')
+        fp = os.path.join(self._root, '{}.npy'.format(str(uuid.uuid4())))
+        np.save(fp, data)
+        resource = self._fs.insert_resource(self.SPEC, fp, resource_kwargs={})
+        datum_id = str(uuid.uuid4())
+        self._fs.insert_datum(resource=resource, datum_id=datum_id,
                               datum_kwargs={})
         return datum_id
 
     def close(self):
-        # make it impossible to accidentally write more datums to this resource
         self._closed = True
-        # close file handle, if applicable
 
     def __enter__(self):
         return self
