+++ redsky/tests/test_streams.py
@@ -1,50 +1,13 @@
-import redsky.event_streams as es
-from ..event_streams import dstar
-from streams.core import Stream
-from ..streamer import StoreSink
-from numpy.testing import assert_allclose
-import numpy as np
-import os
-import uuid
 from functools import partial
 
+import numpy as np
+from numpy.testing import assert_allclose
+from streams.core import Stream
 
-class NpyWriter:
-    """
-    Each call to the ``write`` method saves a file and creates a new filestore
-    resource and datum record.
-    """
-
-    SPEC = 'npy'
-
-    def __init__(self, fs, root):
-        self._root = root
-        self._closed = False
-        self._fs = fs
-        # Open and stash a file handle (e.g., h5py.File) if applicable.
-
-    def write(self, data):
-        """
-        Save data to file, generate and insert new resource and datum.
-        """
-        if self._closed:
-            raise RuntimeError('This writer has been closed.')
-        fp = os.path.join(self._root, '{}.npy'.format(str(uuid.uuid4())))
-        np.save(fp, data)
-        resource = self._fs.insert_resource(self.SPEC, fp, resource_kwargs={})
-        datum_id = str(uuid.uuid4())
-        self._fs.insert_datum(resource=resource, datum_id=datum_id,
-                              datum_kwargs={})
-        return datum_id
-
-    def close(self):
-        self._closed = True
-
-    def __enter__(self):
-        return self
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        self.close()
+import redsky.event_streams as es
+from ..event_streams import dstar
+from ..savers import NpyWriter
+from ..streamer import StoreSink
 
 
 def test_map(exp_db, start_uid1):
